Put token decoder into the right file
